+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Developmental Version
 
+- Reorganized and redesigned all examples
+- Added data sets for exemplification
 - Added grammar for building and handling datasets
 - Rebuilt boxplots, histograms, and data plots
 - Added mean and standard deviation plots